Catch empty responses from Graphite
If a metric/query doesn't exist in Graphite then an empty array will be
returned. Previously this resulted in the exception being returned:

    NoMethodError: undefined method `[]' for nil:NilClass

Replace this with a more user friendly error to indicate that no data has
been returned. This isn't quite the same as not having any valid datapoints,
after dropfirst/droplast.